refactor: migrate `core.event` package to `core.concurrent` to better align with concurrency-related functionality and improve project structure
