+++ core/src/main/java/org/traffichunter/titan/core/channel/ChannelPrimaryIOEventLoopGroup.java
@@ -24,7 +24,7 @@
 package org.traffichunter.titan.core.channel;
 
 import lombok.extern.slf4j.Slf4j;
-import org.traffichunter.titan.core.event.*;
+import org.traffichunter.titan.core.concurrent.*;
 import org.traffichunter.titan.core.util.concurrent.Promise;
 import org.traffichunter.titan.core.util.concurrent.ScheduledPromise;
