+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoop.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
