+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoopBridges.java
@@ -21,29 +21,25 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
-import lombok.extern.slf4j.Slf4j;
+import org.traffichunter.titan.bootstrap.Configurations;
 import org.traffichunter.titan.core.channel.ChannelContext;
-import org.traffichunter.titan.core.channel.ChannelPrimaryIOEventLoop;
-import org.traffichunter.titan.core.channel.ChannelSecondaryIOEventLoop;
-import org.traffichunter.titan.core.util.event.EventLoopConstants;
 
 /**
  * @author yungwang-o
  */
-@Slf4j
-public final class EventLoopFactory {
+public final class EventLoopBridges {
 
-    public static ChannelPrimaryIOEventLoop createPrimaryIOEventLoop() {
-        return new ChannelPrimaryIOEventLoop();
+    private static EventLoopBridge<ChannelContext> INSTANCE = null;
+
+    public static EventLoopBridge<ChannelContext> getInstance() {
+        if(INSTANCE == null) {
+            INSTANCE = new EventLoopBridge<>(Math.max(16, Configurations.taskPendingCapacity()));
+        }
+
+        return INSTANCE;
     }
 
-    public static ChannelSecondaryIOEventLoop createSecondaryIOEventLoop(final int eventLoopNameCount) {
-        return new ChannelSecondaryIOEventLoop(
-                EventLoopConstants.SECONDARY_EVENT_LOOP_THREAD_NAME + "-" + eventLoopNameCount
-        );
-    }
-
-    private EventLoopFactory() {}
+    private EventLoopBridges() { }
 }
