+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoopException.java
@@ -1,4 +1,4 @@
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
 public class EventLoopException extends RuntimeException {
