+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoopFactory.java
@@ -21,12 +21,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
+
+import lombok.extern.slf4j.Slf4j;
+import org.traffichunter.titan.core.channel.ChannelPrimaryIOEventLoop;
+import org.traffichunter.titan.core.channel.ChannelSecondaryIOEventLoop;
+import org.traffichunter.titan.core.util.event.EventLoopConstants;
 
 /**
  * @author yungwang-o
  */
-public interface IOEventLoop extends EventLoop {
+@Slf4j
+public final class EventLoopFactory {
 
-    IOSelector ioHandler();
+    public static ChannelPrimaryIOEventLoop createPrimaryIOEventLoop() {
+        return new ChannelPrimaryIOEventLoop();
+    }
+
+    public static ChannelSecondaryIOEventLoop createSecondaryIOEventLoop(final int eventLoopNameCount) {
+        return new ChannelSecondaryIOEventLoop(
+                EventLoopConstants.SECONDARY_EVENT_LOOP_THREAD_NAME + "-" + eventLoopNameCount
+        );
+    }
+
+    private EventLoopFactory() {}
 }
