+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoopGroup.java
@@ -1,4 +1,4 @@
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
 public interface EventLoopGroup<E extends EventLoop> extends EventLoop {
