+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoopLifeCycle.java
@@ -1,4 +1,4 @@
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
 public interface EventLoopLifeCycle {
