+++ core/src/main/java/org/traffichunter/titan/core/concurrent/EventLoopStatus.java
@@ -1,4 +1,4 @@
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
 import lombok.Getter;
