+++ core/src/main/java/org/traffichunter/titan/core/concurrent/SingleThreadIOEventLoop.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.traffichunter.titan.core.event;
+package org.traffichunter.titan.core.concurrent;
 
 import java.io.IOException;
 import java.nio.channels.SelectionKey;
